Create Download _ remote button
The "Download Remote" button is a convenient and easy-to-use feature that is often used by businesses and individuals who need to access remote resources on a regular basis. By providing a quick and simple way to download and install remote access software, this button makes it easy for users to access the resources they need from anywhere, at any time. 
Read more : https://dimensionsize.com/